check if kubens already cloned

+++ install_kubens_kubectx.sh
@@ -1,11 +1,13 @@
 #!/bin/bash
 
-sudo git clone https://github.com/ahmetb/kubectx /opt/kubectx
-sudo ln -s /opt/kubectx/kubectx /usr/local/bin/kubectx
-sudo ln -s /opt/kubectx/kubens /usr/local/bin/kubens
+if [ ! -e /opt/kubectx ]; then
+  sudo git clone https://github.com/ahmetb/kubectx /opt/kubectx
+  sudo ln -s /opt/kubectx/kubectx /usr/local/bin/kubectx
+  sudo ln -s /opt/kubectx/kubens /usr/local/bin/kubens
 
-sudo apt-get install -y pkg-config
+  sudo apt-get install -y pkg-config
 
-COMPDIR=$(pkg-config --variable=completionsdir bash-completion)
-sudo ln -sf /opt/kubectx/completion/kubens.bash $COMPDIR/kubens
-sudo ln -sf /opt/kubectx/completion/kubectx.bash $COMPDIR/kubectx
+  COMPDIR=$(pkg-config --variable=completionsdir bash-completion)
+  sudo ln -sf /opt/kubectx/completion/kubens.bash $COMPDIR/kubens
+  sudo ln -sf /opt/kubectx/completion/kubectx.bash $COMPDIR/kubectx
+fi
